+++ oldies/openid4vp/utils.py
@@ -1,3 +1,7 @@
+
+from pyeudiw.jwt.utils import decode_jwt_header, decode_jwt_payload
+
+
 def vp_parser(jwt: str) -> Vp:
     """
     Handle the jwt returning the correct VP istance.
@@ -29,6 +33,41 @@ def vp_parser(jwt: str) -> Vp:
         case unsupported:
             raise VPFormatNotSupported(f"parsing of unsupported vp typ [{unsupported}]")
         
+def infer_vp_header_claim(jws: str, claim_name: str) -> Any:
+    """
+    Infer a claim from the header of a VP token.
+
+    :param jws: the VP token
+    :type jws: str
+
+    :param claim_name: the name of the claim to infer
+    :type claim_name: str
+
+    :returns: the value of the claim
+    :rtype: Any
+    """
+    headers = decode_jwt_header(jws)
+    claim_value = headers.get(claim_name, "")
+    return claim_value
+
+
+def infer_vp_payload_claim(jws: str, claim_name: str) -> Any:
+    """
+    Infer a claim from the payload of a VP token.
+
+    :param jws: the VP token
+    :type jws: str
+
+    :param claim_name: the name of the claim to infer
+    :type claim_name: str
+
+    :returns: the value of the claim
+    :rtype: Any
+    """
+    headers = decode_jwt_payload(jws)
+    claim_value: str = headers.get(claim_name, "")
+    return claim_value
+
 
 def infer_vp_typ(jws: str) -> str:
     return infer_vp_header_claim(jws, claim_name="typ")
